Simplifies handling of bindings.
Consistent bindings cannot map a single name to both a C and a t. If
name's value contains a hole, then matching it must have involved the
match function's :hole case (line 56), making the match a proper
decomposition and causing the :name case (line 60) to place a C in the
bindings.