Replaces the ASCII with a define-relation version.
This new version leaves much to be desired in terms of testing. A
match's bindings must be supplied as an input, which is frequently
painful and currently impossible when matching goes through a
non-terminal. It's also unfortunate that you can't say "I expect
exactly these matches" and let the test tell you when there are ones
you didn't consider.