Fixes a bug in in-hole matching
To match t against (in-hole p_1 p_2), we start by matching t against
p_1. If we get the decomposition t = C[t'], then we match t' against
p_2. If matching p_2 uses hole, then we get a decomposition t' =
C'[t''] and therefore report t = C[C'[t'']]. For this match to
ultimately be successful, we need either t'' = hole or the current
in-hole pattern to appear in the first position of another in-hole
pattern whose second position handles t''.